+++ ThirdPartyAdapters/imobile/CHANGELOG.md
@@ -1,5 +1,5 @@
-# i-mobile Adapter for Google Mobile Ads SDK for Android Changelog
+## i-mobile Android Mediation Adapter Changelog
 
-## Version x.y.z.0
-
+#### Version 2.0.20.0
 - Initial release!
+- Added support for banner, interstitial, and native ads.